requirements.txt
falcon===2.0.0
gunicorn==19.9.0
python-dotenv>=0.10.3<0.11.0
cachetools>=3.1.1<3.2.0
numpy==1.17.2
Pillow>=6.1.0<6.2.0
opencv-python>=4.1.1.26<4.1.2.0
black==19.3b0
pytest>=5.1.2<5.2.0
requests
